+++ build.ps1
@@ -1,0 +1,6 @@
+try {
+    vortex build;
+} catch [System.Management.Automation.CommandNotFoundException] {
+    # Fall-back in case Vortex isn't installed on the local machine.
+    dotnet publish "Senpai.sln" "/p:Configuration=Release";
+}