+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [2.0.0-beta.1.0] - 2022-06-15
 ### Added
-* `Argument(string Name, string Description)` - Added a new constructor for `Argument<T>`.
+* `Argument(uint Index, string Name, string Description)` - Added a new constructor for `Argument<T>`.
 * `HelpName` - Added a new property for `Argument<T>`.
 * Support for verbs/subcommands.